Move serial framing into core, temporarily add include path and source to protonc/cpp
