+++ examples/j100/CMakeLists.txt
@@ -34,6 +34,7 @@ add_executable(j100_mcu_c
   j100_mcu.c
   ${GENERATED_FOLDER}/proton__j100_mcu.c
   ${UTILS_FOLDER}/utils.c
+  ${CMAKE_CURRENT_LIST_DIR}/../../core/src/transport/serial.c
 )
 
 target_link_libraries(j100_mcu_c
@@ -47,6 +48,7 @@ target_include_directories(j100_mcu_c
         ${CMAKE_SOURCE_DIR}/nanopb
         ${GENERATED_FOLDER}
         ${UTILS_FOLDER}
+        ${CMAKE_CURRENT_LIST_DIR}/../../core/include
 )
 
 # Install
